pyproject.toml
[tool.poetry]
name = "nonebot-plugin-phigros-insight"
version = "0.0.1"
description = "基于 nonebot2 的支持图片渲染的 phigros 查分插件"
authors = ["shiyihang2007 <467557146@qq.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "nonebot_plugin_phigros_insight"}]
homepage = "https://github.com/shiyihang2007/nonebot-plugin-phigros-insight"
repository = "https://github.com/shiyihang2007/nonebot-plugin-phigros-insight"
documentation = "https://github.com/shiyihang2007/nonebot-plugin-phigros-insight"

# [[tool.poetry.source]]
# name = "tsinghua"
# url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
# default = true

[tool.poetry.dependencies]
python = ">=3.10"
nonebot2 = ">=2.2.0"
nonebot-adapter-onebot = ">=2.4.4"
pillow = ">=10.4.0"
httpx = ">=0.27.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
